refactor: use router
